requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.7
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cortex2==1.0.1
cryptography==35.0.0
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.8
django-cors-headers==3.10.0
django-jazzmin==2.4.8
django-templated-mail==1.1.1
djangochannelsrestframework==0.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-writable-nested==0.6.3
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.3
oauthlib==3.1.1
pandas==1.3.4
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-decouple==3.5
python-dispatch==0.1.31
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
scipy==1.7.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
websocket-client==1.2.1
zope.interface==5.4.0
